psycho/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-environ==0.11.2
psycopg2_binary==2.9.9
requests==2.31.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
djangorestframework-xml==2.0.0
django-solo==2.2.0